src/app/requirements.txt
web.py==0.40
mysql-connector==2.2.9
python-dotenv
